Backend/requirements.txt
flask
flask-cors
spacy
pdfminer.six
scikit-learn